Tighten KIR semantic contract
