+++ resources/foundation/representative.kir.json
@@ -1,6 +1,6 @@
 {
   "metadata": {
-    "kir_schema_version": "0.2",
+    "kir_schema_version": "0.4",
     "name": "Representative Foundation KIR"
   },
   "elements": [
